+++ R/Class-SimParam.R
@@ -1,4 +1,3 @@
-
 #' @title Simulation parameters
 #'
 #' @description